dataflow-job/requirements.txt
apache-beam[gcp]==2.33.0
google-cloud-pubsub==2.7.0
pyyaml==5.4.1
sendgrid